Make all tool tag related functionality conditional on 'enable_tool_tags = True' in the Galaxy configuration file.

+++ lib/galaxy/tools/__init__.py
@@ -130,9 +130,10 @@ class ToolBox( object ):
             key = 'section_' + section.id
             panel_dict[ key ] = section
                 
-        log.info("removing all tool tag associations (" + str( self.sa_session.query( self.app.model.ToolTagAssociation ).count() ) + ")")
-        self.sa_session.query( self.app.model.ToolTagAssociation ).delete()
-        self.sa_session.flush()
+        if self.app.config.get_bool( 'enable_tool_tags', False ):
+            log.info("removing all tool tag associations (" + str( self.sa_session.query( self.app.model.ToolTagAssociation ).count() ) + ")")
+            self.sa_session.query( self.app.model.ToolTagAssociation ).delete()
+            self.sa_session.flush()
         log.info("parsing the tool configuration")
         tree = util.parse_xml( config_filename )
         root = tree.getroot()
